Removed extra components for exams and courses view, Added a route to exams based on name
